Rainbow road to the goal!
For maximum visibility on arbitrary backgrounds, the indicator for your upcoming path is now a changing rainbow-colored line of cells, instead of the hard-to-see slightly-lighter cell from before. This updates jdkgdxds to 0.1.2, and fixes some mistakes in DescriptiveColor (incorrectly returned floats where it should return ints).

+++ squidcore/src/main/java/com/github/yellowstonegames/core/DescriptiveColor.java
@@ -980,7 +980,7 @@ public final class DescriptiveColor {
      * @return the first color this finds that is in-gamut, as if it was moving toward a grayscale color with the same L
      * @see #inGamut(int, int, int) You can use inGamut() if you just want to check whether a color is in-gamut.
      */
-    public static float limitToGamut(int L, int A, int B) {
+    public static int limitToGamut(int L, int A, int B) {
         return limitToGamut(L, A, B, 255);
     }
 
@@ -994,7 +994,7 @@ public final class DescriptiveColor {
      * @return the first color this finds that is in-gamut, as if it was moving toward a grayscale color with the same L
      * @see #inGamut(int, int, int) You can use inGamut() if you just want to check whether a color is in-gamut.
      */
-    public static float limitToGamut(int L, int A, int B, int alpha) {
+    public static int limitToGamut(int L, int A, int B, int alpha) {
         L = Math.min(Math.max(L, 0), 255);
         A = Math.min(Math.max(A, 0), 255);
         B = Math.min(Math.max(B, 0), 255);
